requirements.txt
streamlit>=1.36
python-docx>=1.1.2
reportlab>=4.2
